+++ app/src/main/java/com/uzlov/filmfinder/ui/view/list/films/FilmItemView.kt
@@ -5,5 +5,5 @@ import com.uzlov.filmfinder.ui.view.list.IItemView
 interface FilmItemView: IItemView {
     fun setTitle(text: String)
     fun loadPoster(url: String)
-    fun setRating(countStar: Float)
+    fun setRating(countStar: Double)
 }
